mobile: extract webview overlay out of tabstack and into AuthenticatedApp component, move wer & lure handling to that container
